GPG rule, update tests and comments

+++ Test/Test-RelativePath.ps1
@@ -24,7 +24,7 @@ SOFTWARE.
 #>
 
 #
-# Unit test for adding rules based on computer users
+# Unit test for adding relative path
 #
 
 # Includes
@@ -34,52 +34,36 @@ Import-Module -Name $PSScriptRoot\..\Modules\UserInfo
 Import-Module -Name $PSScriptRoot\..\Modules\ProgramInfo
 Import-Module -Name $PSScriptRoot\..\Modules\FirewallModule
 
-$Group = "Test - Get-UserSDDL"
+#
+# Setup local variables:
+#
+$Group = "Test - Relative path"
 $Profile = "Any"
 
-Write-Host ""
-Write-Host "Remove-NetFirewallRule"
-Write-Host "***************************"
+# Ask user if he wants to load these rules
+Update-Context $IPVersion $Direction $Group
+if (!(Approve-Execute)) { exit }
 
-# Remove previous test
+#
+# TargetProgram installation directories
+#
+$TargetProgramRoot = "C:\Program Files (x86)\Realtek\..\PokerStars.EU"
+
+# Test if installation exists on system
+$global:InstallationStatus = Test-Installation "TargetProgram" ([ref]$TargetProgramRoot) $Terminate
+
+# First remove all existing rules matching group
 Remove-NetFirewallRule -PolicyStore $PolicyStore -Group $Group -Direction $Direction -ErrorAction SilentlyContinue
 
-Write-Host ""
-Write-Host "Get-UserAccounts(Users)"
-Write-Host "***************************"
+#
+# Rules for TargetProgram
+#
 
-[String[]]$UserAccounts = Get-UserAccounts("Users")
-$UserAccounts
-
-Write-Host ""
-Write-Host "Users + Get-UserAccounts(Administrators) + NT SYSTEM"
-Write-Host "***************************"
-
-$UserAccounts = $UserAccounts += (Get-UserAccounts("Administrators"))
-$UserAccounts = $UserAccounts += "NT AUTHORITY\SYSTEM"
-$UserAccounts
-
-Write-Host ""
-Write-Host "Get-UserNames:"
-Write-Host "***************************"
-
-$UserNames = Get-UserNames($UserAccounts)
-$UserNames
-
-Write-Host ""
-Write-Host "Get-UserSDDL:"
-Write-Host "***************************"
-
-$LocalUser = Get-UserSDDL($UserNames)
-$LocalUser
-
-Write-Host ""
-Write-Host "New-NetFirewallRule"
-Write-Host "***************************"
-
+$Program = "$TargetProgramRoot\PokerStars.exe"
+Test-File $Program
 New-NetFirewallRule -Confirm:$Execute -Whatif:$Debug -ErrorAction $OnError -Platform $Platform `
--DisplayName "Get-UserSDDL" -Program Any -Service Any `
--PolicyStore $PolicyStore -Enabled False -Action Allow -Group $Group -Profile $Profile -InterfaceType Any `
--Direction $Direction -Protocol Any -LocalAddress Any -RemoteAddress Any -LocalPort Any -RemotePort Any `
--LocalUser $LocalUser `
+-DisplayName "TargetProgram" -Service Any -Program $Program `
+-PolicyStore $PolicyStore -Enabled True -Action Allow -Group $Group -Profile $Profile -InterfaceType $Interface `
+-Direction $Direction -Protocol TCP -LocalAddress Any -RemoteAddress Internet4 -LocalPort Any -RemotePort 80, 443, 26002 `
+-LocalUser $UserAccountsSDDL `
 -Description ""
